+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-5928] Allow to publish deprecated concepts
+
 ## [5.12.0] 2023-08-16
 
 ### Added